+++ src/main/java/com/abdala/demo/repository/ArticleRepo.java
@@ -11,15 +11,15 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface ArticleRepo extends JpaRepository<Article, Long> {
+public interface ArticleRepo extends JpaRepository<Article, Integer> {
 
     @Query("SELECT a FROM Article a  WHERE a.tagName = :tagName")
     List<Article> findArticlesByTagName(@Param("tagName") String tagName);
 
   @Query("select a from Article a join User x where x.id=:authorId")
-    List<Article> findByAuthorId(Long authorId);
+    List<Article> findByAuthorId(Integer authorId);
 
-    List<Article> findFavoriteArticlesByUserId(Long userId);
+    List<Article> findFavoriteArticlesByUserId(Integer userId);
 
 
 }
