+++ src/main/java/com/abdala/demo/repository/UserRepo.java
@@ -10,11 +10,11 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface UserRepo extends JpaRepository<User,Long> {
+public interface UserRepo extends JpaRepository<User,Integer> {
 
-    Optional<User> findById(Long id);
+    Optional<User> findById(Integer id);
 
-    List<User> findFollowedUsersByUserId(Long userId);
+    List<User> findFollowedUsersByUserId(Integer userId);
 
     Optional<User> findByEmail(String email);
